Added partisan lean script
