+++ .env.local
@@ -1,0 +1,1 @@
+MONGODB_URL="mongodb://localhost:27017"
