+++ .env.original
@@ -1,2 +1,1 @@
-MONGODB_USER=
-MONGODB_PASSWORD=
+MONGODB_URL=
