+++ python/requirements.txt
@@ -1,3 +1,4 @@
 pandas
 requests
 pymongo
+dotenv
